+++ efm32/1.0.1/libraries/EmusDevices/MLX90614Device.h
@@ -32,7 +32,7 @@ class MLX90614Device : public Device
         virtual float temperature();
          
         virtual Device::ValueStruct readValue(int index);
-        virtual float numValues(); 
+        virtual uint32_t numValues(); 
         
    protected:   
         TwoWire *myWire;
