+++ efm32/1.0.1/libraries/EmusDevices/MPL3115A2Device.h
@@ -38,7 +38,7 @@ class MPL3115A2Device : public Device
 
          
         virtual Device::ValueStruct readValue(int index);
-        virtual float numValues(); 
+        virtual uint32_t numValues(); 
         
     protected:
