+++ efm32/1.0.1/libraries/EmusDevices/TCS34725Device.cpp
@@ -374,7 +374,7 @@ Device::ValueStruct TCS34725Device::readValue(int index)
     return output;
 }
 
-float TCS34725Device::numValues()
+uint32_t TCS34725Device::numValues()
 {
     return 6; 
 }
